LPS-89154 Moves code from type controller, no logic changes
