+++ modules/apps/layout/layout-content-page-editor-web/src/main/java/com/liferay/layout/content/page/editor/web/internal/listener/LayoutModelListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.type.controller.content.internal.listener;
+package com.liferay.layout.content.page.editor.web.internal.listener;
 
 import com.liferay.exportimport.kernel.lar.ExportImportThreadLocal;
 import com.liferay.fragment.model.FragmentEntryLink;
