+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/init.jsp
@@ -22,9 +22,9 @@ taglib uri="http://liferay.com/tld/soy" prefix="soy" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
-<%@ page import="com.liferay.layout.type.controller.content.internal.constants.ContentLayoutPortletKeys" %><%@
+<%@ page import="com.liferay.layout.content.page.editor.web.internal.constants.ContentLayoutPortletKeys" %><%@
+page import="com.liferay.layout.content.page.editor.web.internal.display.context.FragmentsEditorDisplayContext" %><%@
 page import="com.liferay.layout.type.controller.content.internal.constants.ContentLayoutTypeControllerWebKeys" %><%@
-page import="com.liferay.layout.type.controller.content.internal.display.context.FragmentsEditorDisplayContext" %><%@
 page import="com.liferay.portal.kernel.json.JSONFactoryUtil" %><%@
 page import="com.liferay.portal.kernel.json.JSONSerializer" %><%@
 page import="com.liferay.portal.kernel.util.PortalUtil" %>
