+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/view.jsp
