+++ modules/apps/layout/layout-type-controller/layout-type-controller-content/src/main/resources/META-INF/resources/layout_editor_toolbar/init.jsp
@@ -16,4 +16,4 @@
 
 <%@ taglib uri="http://liferay.com/tld/portlet" prefix="liferay-portlet" %>
 
-<%@ page import="com.liferay.layout.type.controller.content.internal.constants.ContentLayoutPortletKeys" %>
+<%@ page import="com.liferay.layout.content.page.editor.web.internal.constants.ContentLayoutPortletKeys" %>
